restructuring downloading and processing code to work for various machines and signals flexibly
